Implement OS specific tests
