+++ internal/config/config_windows_test.go
@@ -1,0 +1,9 @@
+package config_test
+
+import (
+	"testing"
+)
+
+func TestCreate(t *testing.T) {
+	t.Error("Was executed")
+}
